Extract update tables into own class
This was a bit unruly as a module. Extracting into own class provides better encapsulation.
